create edit profile page component
